+++ frontend/src/components/PairRoomOnboarding/PairRoomSettingSection/PairRoomSettingSection.styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Layout = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 8rem;
+`;
+
+export const ButtonWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+
+  width: 100%;
+  margin-top: 5rem;
+`;
